Move create_identity_vector() and create_identity_matrix() function from generalized_alpha_time_integrator-> utilities
