+++ src/rigid_pendulum_poc/utilities.h
@@ -1,6 +1,11 @@
 #pragma once
 
+#include <Kokkos_Core.hpp>
+
 namespace openturbine::rigid_pendulum {
+
+using HostView1D = Kokkos::View<double*, Kokkos::HostSpace>;
+using HostView2D = Kokkos::View<double**, Kokkos::HostSpace>;
 
 // TODO: Move the following definitions to a constants.h file in a common math directory
 static constexpr double kTOLERANCE = 1e-6;
@@ -20,5 +25,17 @@ bool close_to(double a, double b, double epsilon = kTOLERANCE);
  * @param  angle: Angle to be wrapped, in radians
  */
 double wrap_angle_to_pi(double angle);
+
+/*!
+ * @brief  Creates an identity vector of size (size x 1)
+ * @param  size: Size of the identity vector
+ */
+HostView1D create_identity_vector(size_t size);
+
+/*!
+ * @brief  Creates an identity matrix of size (size x size)
+ * @param  size: Size of the identity matrix
+ */
+HostView2D create_identity_matrix(size_t size);
 
 }  // namespace openturbine::rigid_pendulum
